:recycle: Cleanup TypeScript
